feat!: drop old schema migrations
